ajustes banco de dados

+++ database/migrations/2026_03_14_132633_create_movimentacao_produtos_table.php
@@ -11,28 +11,30 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('movimentacao_produtos', function (Blueprint $table) {
-            $table->id();
-            
-            $table->unsignedBigInteger('movimentacao_id');
-            $table->foreign('movimentacao_id')->references('id')->on('movimentacoes')->onDelete('cascade');
-            
-            $table->unsignedBigInteger('produto_id');
-            $table->foreign('produto_id')->references('id')->on('produtos');
+        if (!Schema::hasTable('movimentacao_produtos')) {
+            Schema::create('movimentacao_produtos', function (Blueprint $table) {
+                $table->id();
+                
+                $table->unsignedBigInteger('movimentacao_id');
+                $table->foreign('movimentacao_id')->references('id')->on('movimentacoes')->onDelete('cascade');
+                
+                $table->unsignedBigInteger('produto_id');
+                $table->foreign('produto_id')->references('id')->on('produtos');
 
-            $table->unsignedBigInteger('fornecedor_id')->nullable();
-            $table->foreign('fornecedor_id')->references('id')->on('fornecedores');
-            
-            $table->integer('quantidade');
-            $table->double('valor_unitario')->nullable();
-            $table->double('valor_total')->nullable();
-            
-            $table->string('endereco')->nullable();
-            $table->string('corredor')->nullable();
-            $table->string('posicao')->nullable();
-            
-            $table->timestamps();
-        });
+                $table->unsignedBigInteger('fornecedor_id')->nullable();
+                $table->foreign('fornecedor_id')->references('id')->on('fornecedores');
+                
+                $table->integer('quantidade');
+                $table->double('valor_unitario')->nullable();
+                $table->double('valor_total')->nullable();
+                
+                $table->string('endereco')->nullable();
+                $table->string('corredor')->nullable();
+                $table->string('posicao')->nullable();
+                
+                $table->timestamps();
+            });
+        }
     }
 
     /**
